requirements.txt
flask
Flask-APScheduler
flask-sqlalchemy
flask-compress
flask-bcrypt
flask-admin
flask-login
flask-mail
flask-wtf
python-dotenv
itsdangerous
apscheduler
sqlalchemy
requests
jikanpy
wtforms
pathlib
Pillow
regex
pytz